+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 15.1.0
-- Metrics: output range squeezed to [0, 1]
+- MAPE metric: output range squeezed to [0, 1]
 
 ## 15.0.1
 - RegressorAssessor: unit tests added